requirements.txt
fastapi==0.115.11
langchain==0.3.20
langchain_core==0.3.42
langchain_openai==0.3.8
langgraph==0.3.5
openai==1.65.4
pandas==2.2.3
pydantic==2.10.6
pymongo==4.11.2
python-dotenv==1.0.1
supabase==2.13.0
tiktoken==0.9.0
uvicorn==0.15.0
